Refactor index.tsx to be in app router
